MapquestAPI
This is my version of the capstone project my team and I worked on.
This is my branch as I need something that wasn't copy written (aka the
final project for the stakeholder) and contained mostly my code and my
back-end partner's code.  Also I redesigned the interface so legal
reasons.  This program is written in C# and it uses Visual Studio's 2013
IDE.  For any more questions pertaining to specifics, please refer to
the documentation and the uml files.